+++ src/frontend/src/tables/settings/TemplateTable.tsx
@@ -5,11 +5,9 @@ import { useCallback, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { AddItemButton } from '../../components/buttons/AddItemButton';
-import {
-  CodeEditor,
-  PreviewArea,
-  TemplateEditor
-} from '../../components/editors/TemplateEditor';
+import { TemplateEditor } from '../../components/editors/TemplateEditor';
+import { CodeEditor } from '../../components/editors/TemplateEditor/CodeEditor';
+import { PreviewArea } from '../../components/editors/TemplateEditor/PDFPreview';
 import {
   ActionDropdown,
   DeleteItemAction,
